move math funcs to new parser to have <70 methods in builtinfuncparser
