+++ modules/engine/src/main/scala/com/gsk/kg/engine/ExpressionF.scala
@@ -207,11 +207,11 @@ object ExpressionF {
       case ConditionOrder.ASC(e)                       => ASC(e)
       case ConditionOrder.DESC(e)                      => DESC(e)
       case BuiltInFunc.UUID()                          => UUID()
-      case BuiltInFunc.CEIL(s)                         => CEIL(s)
-      case BuiltInFunc.ROUND(s)                        => ROUND(s)
-      case BuiltInFunc.RAND()                          => RAND()
-      case BuiltInFunc.ABS(s)                          => ABS(s)
-      case BuiltInFunc.FLOOR(s)                        => FLOOR(s)
+      case MathFunc.CEIL(s)                            => CEIL(s)
+      case MathFunc.ROUND(s)                           => ROUND(s)
+      case MathFunc.RAND()                             => RAND()
+      case MathFunc.ABS(s)                             => ABS(s)
+      case MathFunc.FLOOR(s)                           => FLOOR(s)
       case BuiltInFunc.STRUUID()                       => STRUUID()
       case DateTimeFunc.NOW()                          => NOW()
       case DateTimeFunc.YEAR(s)                        => YEAR(s)
@@ -340,11 +340,11 @@ object ExpressionF {
       case ASC(e)                     => ConditionOrder.ASC(e)
       case DESC(e)                    => ConditionOrder.DESC(e)
       case UUID()                     => BuiltInFunc.UUID()
-      case CEIL(s)                    => BuiltInFunc.CEIL(s)
-      case ROUND(s)                   => BuiltInFunc.ROUND(s)
-      case RAND()                     => BuiltInFunc.RAND()
-      case ABS(s)                     => BuiltInFunc.ABS(s)
-      case FLOOR(s)                   => BuiltInFunc.FLOOR(s)
+      case CEIL(s)                    => MathFunc.CEIL(s)
+      case ROUND(s)                   => MathFunc.ROUND(s)
+      case RAND()                     => MathFunc.RAND()
+      case ABS(s)                     => MathFunc.ABS(s)
+      case FLOOR(s)                   => MathFunc.FLOOR(s)
       case STRUUID()                  => BuiltInFunc.STRUUID()
       case NOW()                      => DateTimeFunc.NOW()
       case YEAR(s)                    => DateTimeFunc.YEAR(s)
